requirements.txt
fastapi[all]
sqlmodel
psycopg2
requests
pytest
flake8
pyjwt
passlib[bcrypt]
mypy